documentation user and partial expanse
